requirements.txt
Django==3.0.8
djangorestframework==3.11.0
Pillow==7.2.0
psycopg2-binary==2.8.5
twilio==6.45.0
requests==2.24.0
google==2.0.2
gunicorn==20.0.4
